Finished the MVP.
Added endpoint to calculate and return the final rating.
Currently I'm hardcoding the Id of the book, but it's fine for this demo.

Using query operation instead of scan for getting reviews.
This allows me to limit the number of results dispayed and paginate.

Adding a hardcoded ID for all reviews, when posting.
Just to make things simple.

Added favicon for final polish.